combined bore encoder with ArmSubsystem
