+++ src/main/java/frc/robot/Robot.java
@@ -29,7 +29,7 @@ public class Robot extends TimedRobot {
   public static final String VISION = "Vision";
 
   // change this to match the subsystem container you want to use, or GAME for complete robot
-  public static final String container = DRIVE;
+  public static final String container = ARM;
 
   private Command m_autonomousCommand;
