+++ src/main/java/frc/robot/commands/ArmCommand.java
@@ -1,5 +1,7 @@
 package frc.robot.commands;
 
 public class ArmCommand {
-    //setArmPosition(start, goal), velocity
+    
+    //setArmPosition(startAngle, goalAngle), velocity
+
 }
