--- src/main/java/frc/robot/subsystems/Arm.java
@@ -1,25 +1,0 @@
-package frc.robot.subsystems;
-
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
-
-
-import com.ctre.phoenix.motorcontrol.ControlMode;
-import com.ctre.phoenix.motorcontrol.NeutralMode;
-//
-import com.ctre.phoenix6.signals.NeutralModeValue;
-import com.ctre.phoenix6.configs.MotorOutputConfigs;
-//
-
-import com.ctre.phoenix6.controls.MotionMagicVelocityTorqueCurrentFOC;
-import com.ctre.phoenix6.hardware.TalonFX;
-
-//import org.littletonrobotics.junction.AutoLog;
-
-public class Arm extends SubsystemBase {
-  //Bore encoder, Talonfx motor, Magic motion?, 
-  //Functions: getCurrentPosition/getCurrentAngle, setTargetPosition, getTorqueVelocity, setVelocity, 
-  //
-
-  
-}
